added test to confirm chat history, cache and rag work together + reorganized tests #15
