+++ src/tests/unit-tests/endpoint-agent.unit.test.ts
@@ -1,15 +1,15 @@
 import {
   defineChatEndpoint,
   getChatEndpointRunner,
-} from "../endpoints/endpoints";
-import { setupGenkit } from "../genkit/genkit";
+} from "../../endpoints/endpoints";
+import { setupGenkit } from "../../genkit/genkit";
 
 /**
- * Test suite for Chat Endpoint Basic Functionality.
+ * Test suite for testing chat agent config for a chat endpoint.
  *
  * Some tests include the use of LLM model, defining a chat agent, defining API key store, defining chat history store, and defining cache store.
  */
-describe("Test - Chat Endpoint Core Funtionality Tests", () => {
+describe("Test - Chat Endpoint Agent Config Tests", () => {
   // Initialize endpoint runner
   const runEndpoint = getChatEndpointRunner();
 
@@ -28,17 +28,28 @@ describe("Test - Chat Endpoint Core Funtionality Tests", () => {
   const defaultTimeout = 10000; // 10 secondss
 
   if (Tests.define_chat_endpoint)
-    test("Define chat endpoint", () => {
-      const endpoint = defineChatEndpoint({ endpoint: "test-chat" });
+    test("Define chat endpoint with chat agent config", () => {
+      const endpoint = defineChatEndpoint({
+        endpoint: "test-chat-agent",
+        chatAgentConfig: {
+          model: "gemini10Pro",
+        },
+      });
       expect(endpoint).toBeDefined();
     });
 
   if (Tests.confirm_response_generation)
     test(
-      "Confirm response generation",
+      "Confirm response generation for endpoint with chat agent config",
       async () => {
         const endpoint = defineChatEndpoint({
-          endpoint: "test-chat-open-response",
+          endpoint: "test-chat-agent-response",
+          chatAgentConfig: {
+            model: "gemini10Pro",
+            modelConfig: {
+              temperature: 0.9,
+            },
+          },
         });
         const response = await runEndpoint(endpoint, {
           query: "How can you help? In one sentence.",
