+++ src/tests/unit-tests/endpoint-close-ended.unit.test.ts
@@ -1,15 +1,15 @@
 import {
   defineChatEndpoint,
   getChatEndpointRunner,
-} from "../endpoints/endpoints";
-import { setupGenkit } from "../genkit/genkit";
+} from "../../endpoints/endpoints";
+import { setupGenkit } from "../../genkit/genkit";
 
 /**
- * Test suite for testing chat agent config for a chat endpoint.
+ * Test suite for Close-ended Chat Endpoint.
  *
  * Some tests include the use of LLM model, defining a chat agent, defining API key store, defining chat history store, and defining cache store.
  */
-describe("Test - Chat Endpoint Agent Config Tests", () => {
+describe("Test - Close-ended Chat Endpoint Tests", () => {
   // Initialize endpoint runner
   const runEndpoint = getChatEndpointRunner();
 
@@ -20,36 +20,31 @@ describe("Test - Chat Endpoint Agent Config Tests", () => {
   // Tests to be performed
   // Set to true to run the test
   const Tests = {
-    define_chat_endpoint: true,
+    define_close_ended_chat: true,
     confirm_response_generation: true,
   };
 
   // default test timeout
   const defaultTimeout = 10000; // 10 secondss
 
-  if (Tests.define_chat_endpoint)
-    test("Define chat endpoint with chat agent config", () => {
+  if (Tests.define_close_ended_chat)
+    test("Define chat endpoint", () => {
       const endpoint = defineChatEndpoint({
-        endpoint: "test-chat-agent",
-        chatAgentConfig: {
-          model: "gemini10Pro",
-        },
+        endpoint: "test-chat-close",
+        agentType: "close-ended",
+        topic: "Firebase",
       });
       expect(endpoint).toBeDefined();
     });
 
   if (Tests.confirm_response_generation)
     test(
-      "Confirm response generation for endpoint with chat agent config",
+      "Confirm response generation",
       async () => {
         const endpoint = defineChatEndpoint({
-          endpoint: "test-chat-agent-response",
-          chatAgentConfig: {
-            model: "gemini10Pro",
-            modelConfig: {
-              temperature: 0.9,
-            },
-          },
+          endpoint: "test-chat-close-response",
+          agentType: "close-ended",
+          topic: "Firebase",
         });
         const response = await runEndpoint(endpoint, {
           query: "How can you help? In one sentence.",
